Get rid of unnecessary date helper
No longer serves a useful purpose
