--- dmscripts/helpers/date_helpers.py
@@ -1,14 +1,0 @@
-from datetime import datetime, timedelta
-from dmutils.formats import DATE_FORMAT
-
-
-def get_date_closed(date_closed):
-    """
-    Defaults to yesterday's date (used for notifying buyers when their briefs close).
-    :param date_closed: YYYY-MM-DD string supplied via script argument
-    :return: Python datetime object
-    """
-    if date_closed is None:
-        return (datetime.utcnow() - timedelta(days=1)).date()
-    else:
-        return datetime.strptime(date_closed, DATE_FORMAT).date()
